crossdock - let client register its behaviors itself
