+++ crossdock/client/random/rand.go
